requirements.txt
bcrypt
mysql-connector-python
python-dotenv
